Tons of bugfixes (small) and template changes. Aurora template looks nice for now. (even with the nav color)

+++ src/Lotgd/ExpireChars.php
@@ -50,9 +50,9 @@ class ExpireChars
      */
     private static function cleanupExpiredAccounts(): void
     {
-        $old = getsetting('expireoldacct', 45);
-        $new = getsetting('expirenewacct', 10);
-        $trash = getsetting('expiretrashacct', 1);
+        $old = (int)getsetting('expireoldacct', 45);
+        $new = (int)getsetting('expirenewacct', 10);
+        $trash = (int)getsetting('expiretrashacct', 1);
 
         $rows = self::fetchAccountsToExpire($old, $new, $trash);
         if (empty($rows)) {
@@ -141,7 +141,7 @@ class ExpireChars
      */
     private static function notifyUpcomingExpirations(): void
     {
-        $old = max(1, getsetting('expireoldacct', 45) - getsetting('notifydaysbeforedeletion', 5));
+        $old = max(1, ((int)getsetting('expireoldacct', 45) - ((int)getsetting('notifydaysbeforedeletion', 5)));
         $sql = 'SELECT login,acctid,emailaddress FROM ' . db_prefix('accounts') .
             " WHERE 1=0 " . ($old > 0 ? "OR (laston < '" . date('Y-m-d H:i:s', strtotime("-$old days")) . "')" : '') .
             " AND emailaddress!='' AND sentnotice=0 AND (superuser&" . NO_ACCOUNT_EXPIRATION . ')=0';
